+++ src/textpath.js
@@ -3,32 +3,14 @@ SVG.TextPath = SVG.invent({
   create: 'textPath'
 
   // Inherit from
-, inherit: SVG.Parent
+, inherit: SVG.Text
 
   // Define parent class
-, parent: SVG.Text
+, parent: SVG.Parent
 
   // Add parent method
-, construct: {
+, extend: {
     morphArray: SVG.PathArray
-    // Create path for text to run on
-  , path: function(d) {
-      // create textPath element
-      var path  = new SVG.TextPath
-        , track = this.doc().defs().path(d)
-
-      // move lines to textpath
-      while (this.node.hasChildNodes())
-        path.node.appendChild(this.node.firstChild)
-
-      // add textPath element as child node
-      this.node.appendChild(path.node)
-
-      // link textPath to path and add content
-      path.attr('href', '#' + track, SVG.xlink)
-
-      return this
-    }
     // return the array of the path track element
   , array: function() {
       var track = this.track()
@@ -46,17 +28,50 @@ SVG.TextPath = SVG.invent({
 
       return (d == null) ? pathArray : this
     }
-    // Get the path track element
+    // Get the path element
   , track: function() {
-      var path = this.textPath()
-
-      if (path)
-        return path.reference('href')
-    }
-    // Get the textPath child
-  , textPath: function() {
-      if (this.node.firstChild && this.node.firstChild.nodeName == 'textPath')
-        return SVG.adopt(this.node.firstChild)
+      return this.reference('href')
     }
   }
+, construct: {
+    textPath: function(text, path) {
+      return this.defs().path(path).text(text).addTo(this)
+    }
+  }
+})
+
+SVG.extend([SVG.Text], {
+    // Create path for text to run on
+  path: function(track) {
+    var path = new SVG.TextPath
+
+    // if d is a path, reuse it
+    if(!(track instanceof SVG.Path)) {
+      // create path element
+      track = this.doc().defs().path(track)
+    }
+
+    // link textPath to path and add content
+    path.attr('href', '#' + track, SVG.xlink)
+
+    // add textPath element as child node and return textPath
+    return this.put(path)
+  }
+  // Todo: make this plural?
+  // Get the textPath children
+  , textPath: function() {
+    return this.select('textPath')
+  }
+})
+
+SVG.extend([SVG.Path], {
+  // creates a textPath from this path
+  text: function(text) {
+    if(text instanceof SVG.Text) {
+      var txt = text.text()
+      return text.clear().path(this).text(txt)
+    }
+    return this.parent().put(new SVG.Text()).path(this).text(text)
+  }
+  // TODO: Maybe add `targets` to get all textPaths associated with this path
 })